Make BlobSchedule its own class
