+++ config/src/main/java/org/hyperledger/besu/config/BlobScheduleOptions.java
@@ -89,77 +89,4 @@ public class BlobScheduleOptions {
     getFutureEips().ifPresent(bs -> builder.put(FUTURE_EIPS_KEY, bs.asMap()));
     return builder.build();
   }
-
-  /** The Blob schedule for a particular fork. */
-  public static class BlobSchedule {
-    private final int target;
-    private final int max;
-    private final int baseFeeUpdateFraction;
-
-    /** The constant PRAGUE_DEFAULT. */
-    public static final BlobSchedule NO_BLOBS = new BlobSchedule(0, 0, 3338477);
-
-    /** The constant CANCUN_DEFAULT. */
-    public static final BlobSchedule CANCUN_DEFAULT = new BlobSchedule(3, 6, 3338477);
-
-    /** The constant PRAGUE_DEFAULT. */
-    public static final BlobSchedule PRAGUE_DEFAULT = new BlobSchedule(6, 9, 5007716);
-
-    /** The constant OSAKA_DEFAULT. */
-    public static final BlobSchedule OSAKA_DEFAULT = new BlobSchedule(9, 12, 5007716);
-
-    /**
-     * Instantiates a new Blob schedule.
-     *
-     * @param blobScheduleConfigRoot the blob schedule config root
-     */
-    public BlobSchedule(final ObjectNode blobScheduleConfigRoot) {
-      this.target = JsonUtil.getInt(blobScheduleConfigRoot, "target").orElseThrow();
-      this.max = JsonUtil.getInt(blobScheduleConfigRoot, "max").orElseThrow();
-      this.baseFeeUpdateFraction =
-          JsonUtil.getInt(blobScheduleConfigRoot, "basefeeupdatefraction").orElseThrow();
-    }
-
-    private BlobSchedule(final int target, final int max, final int baseFeeUpdateFraction) {
-      this.target = target;
-      this.max = max;
-      this.baseFeeUpdateFraction = baseFeeUpdateFraction;
-    }
-
-    /**
-     * Gets target.
-     *
-     * @return the target
-     */
-    public int getTarget() {
-      return target;
-    }
-
-    /**
-     * Gets max.
-     *
-     * @return the max
-     */
-    public int getMax() {
-      return max;
-    }
-
-    /**
-     * Gets base fee update fraction.
-     *
-     * @return the base fee update fraction
-     */
-    public int getBaseFeeUpdateFraction() {
-      return baseFeeUpdateFraction;
-    }
-
-    /**
-     * As map.
-     *
-     * @return the map
-     */
-    Map<String, Object> asMap() {
-      return Map.of("target", target, "max", max, "baseFeeUpdateFraction", baseFeeUpdateFraction);
-    }
-  }
 }
